src/python/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.5
idna==3.3
PyJWT==2.4.0
PyMySQL==1.0.2
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.9
